Diff viewer prototype
Big mess right now, but working
