+++ FModel/ViewModels/Commands/LoadCommand.cs
@@ -60,7 +60,7 @@ public class LoadCommand : ViewModelCommand<LoadingModesViewModel>
 
         await Task.WhenAll(
             _applicationView.CUE4Parse.LoadLocalizedResources(), // load locres if not already loaded,
-            _applicationView.CUE4Parse.LoadVirtualPaths(), // load virtual paths if not already loaded
+            _applicationView.CUE4Parse.LoadAllVirtualPaths(), // load virtual paths if not already loaded
             _threadWorkerView.Begin(cancellationToken =>
             {
                 // filter what to show
